plot_worlds_exporter/requirements.txt
anvil-parser @ git+https://github.com/MinecraftContest/anvil-parser.git
